% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize.R
\name{dmeanSampleSize}
\alias{dmeanSampleSize}
\title{Difference in means sample size}
\usage{
dmeanSampleSize(moe, sd, conf = 0.95)
}
\arguments{
\item{moe}{margin of error expressed in your measure's units}

\item{sd}{the common standard deviation of the two populations expressed in your measure's units}

\item{conf}{confidence level associated with margin of error (default = 0.95)}
}
\value{
Estimated sample size for each group.
}
\description{
Calculate sample size (per group) needed to estimate a difference in means within a
given margin of error. This function assumes the means come from
Normal distributions with common variances. Those are big and likely untrue
assumptions, therefore the estimated sample size should be treated as a rough estimate.
}
\examples{
## Approximate sample size needed to estimate the difference in means between two
## populations within 1.5 units, each with standard deviation 20 units.
dmeanSampleSize(moe = 1.5, sd = 20)

}
\references{
Hogg, R. V. and Tanis, E. A. (2006), \emph{Probability and Statistical Inference 7e}, Pearson, New Jersey.
}

