% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samplesize.R
\name{finitePropSampleSize}
\alias{finitePropSampleSize}
\title{proportion sample size - finite (known) population}
\usage{
finitePropSampleSize(N, moe, p = 0.5, conf = 0.95)
}
\arguments{
\item{N}{population size}

\item{moe}{margin of error}

\item{p}{estimated proportion (default = 0.5)}

\item{conf}{confidence level (default = 0.95)}
}
\value{
Estimated sample size needed to estimate a proportion within a given
  margin of error for a finite (known) population size.
}
\description{
Calculate sample size needed to estimate a proportion within a given margin
of error for a finite (known) population size.
}
\examples{
finitePropSampleSize(N=3000, moe=0.03)
}

