% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_and_c.R
\name{propSampCutoff}
\alias{propSampCutoff}
\title{Proportion sample size and cutoff}
\usage{
propSampCutoff(alpha, beta, null, alt, direction = c("greater", "less"))
}
\arguments{
\item{alpha}{Type I error}

\item{beta}{Type II error}

\item{null}{null proportion}

\item{alt}{alternative proportion}

\item{direction}{"greater" (default) or "less"}
}
\value{
Cutoff (critical region boundary) and sample size given type I error,
type II error, null proportion, and alternative proportion
}
\description{
Given type I error, type II error, null proportion, and
alternative proportion, return cutoff (critical region boundary) and sample size.
}
\examples{
# example 9.1-3
propSampCutoff(alpha = 0.05,beta = 0.1,null = 0.5,alt = 0.25,
direction = "less")
# problem 9.1-11
propSampCutoff(alpha = 0.05,beta = 0.1,null = 1/26,alt = 1/10)

}
\references{
Hogg, R. V. and Tanis, E. A. (2006), \emph{Probability and Statistical Inference 7e},
Pearson, New Jersey. Chapter 9.1.
}

