% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samplesize.R
\name{propSampleSize}
\alias{propSampleSize}
\title{proportion sample size}
\usage{
propSampleSize(moe, p = 0.5, conf = 0.95)
}
\arguments{
\item{moe}{margin of error}

\item{p}{estimated proportion (default = 0.5)}

\item{conf}{confidence level (default = 0.95)}
}
\value{
Estimated sample size needed to estimate a proportion within a given
  margin of error.
}
\description{
Calculate sample size needed to estimate a proportion within a given margin
of error.
}
\examples{
propSampleSize(moe=0.03)
}

