% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize.R
\name{dpropSampleSize}
\alias{dpropSampleSize}
\title{difference in proportions sample size}
\usage{
dpropSampleSize(moe, p1 = 0.5, p2 = 0.5, conf = 0.95)
}
\arguments{
\item{moe}{margin of error}

\item{p1}{estimated proportion in group 1 (default = 0.5)}

\item{p2}{estimated proportion in group 2 (default = 0.5)}

\item{conf}{confidence level (default = 0.95)}
}
\value{
Estimated sample size needed to estimate a difference in proportions
  within a given margin of error. The sample size is for each group.
}
\description{
difference in proportions sample size
}
\examples{
dpropSampleSize(moe = 0.02, p1 = 0.11, p2 = 0.09)

}
\references{
Hogg, R. V. and Tanis, E. A. (2006), \emph{Probability and Statistical Inference 7e}, Pearson, New Jersey.
}

