% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize.R
\name{propSampleSize}
\alias{propSampleSize}
\title{Proportion sample size}
\usage{
propSampleSize(moe, p = 0.5, conf = 0.95)
}
\arguments{
\item{moe}{margin of error expressed as a proportion (0 < moe < 1)}

\item{p}{estimated proportion (default = 0.5); leave as is for most conservative estimate.}

\item{conf}{confidence level associated with margin of error (default = 0.95)}
}
\value{
Estimated sample size needed to estimate a proportion within a given
  margin of error.
}
\description{
Calculate sample size needed to estimate a proportion within a given margin
of error.
}
\examples{
propSampleSize(moe=0.03)

}
\references{
Hogg, R. V. and Tanis, E. A. (2006), \emph{Probability and Statistical Inference 7e}, Pearson, New Jersey.
}

